:recycle: [refactory]: Update Directory
- Update directory (collection of pages.)
