+++ src/pages/collections/application/context/atom.ts
@@ -1,6 +1,6 @@
 import { atom } from 'recoil';
-import Https from '../../../service/https';
-import CollectionApi from '../domain/collection';
+import Https from '../../../../service/https';
+import CollectionApi from '../../application/domain/collection';
 
 const baseURL = process.env.REACT_APP_BASE_URL!;
 const httpService = new Https(baseURL);
