+++ src/pages/collections/application/custom_hooks/index.ts
@@ -1,0 +1,3 @@
+export { useCollectionFetch } from './useCollectionFetch';
+export { useCollectionDel } from './useCollectionDel';
+export { useCollectionsDel } from './useCollectionsDel';
