+++ src/pages/collections/application/custom_hooks/useCollectionsDel.ts
@@ -3,8 +3,8 @@ import { useRecoilValue } from 'recoil';
 import { AxiosResponse } from 'axios';
 
 import { thisApi } from '../context/atom';
-import { toastStyle } from '../../../utils/toastConfig';
-import { queryKey } from '../../../utils/config';
+import { toastStyle } from '../../../../utils/toastConfig';
+import { queryKey } from '../../../../utils/config';
 
 interface ICollectionsDel {
   delsLoading: boolean;
