+++ src/pages/collections/application/dto/index.ts
@@ -1,0 +1,1 @@
+export * as dto from './dto';
