+++ src/pages/collections/page/CollectionBody/CollectionControl.tsx
@@ -1,7 +1,7 @@
 import styled from 'styled-components';
 
-import { colors } from '../../../utils/config';
-import { media } from '../../../utils/mixins';
+import { colors } from '../../../../utils/config';
+import { media } from '../../../../utils/mixins';
 
 interface ICollectionControl {
   total: number;
