+++ src/pages/collections/page/CollectionBody/CollectionList/index.ts
@@ -1,0 +1,1 @@
+export { default as CollectionList } from './CollectionList';
