+++ src/pages/collections/page/CollectionBody/index.ts
@@ -1,0 +1,1 @@
+export { default as CollectionBody } from './CollectionBody';
