+++ src/pages/collections/page/index.ts
@@ -1,0 +1,1 @@
+export { default as Collection } from './Collection';
