+++ src/router/MainRouter.tsx
@@ -1,7 +1,7 @@
 import { BrowserRouter, Route, Switch } from 'react-router-dom';
 
-import Collections from '../pages/collections/application';
-import Users from '../pages/users/application';
+import Collections from '../pages/collections/page/Collection';
+//import Users from '../pages/users/application';
 import Notfound from '../components/Notfound';
 
 function MainRouter() {
@@ -10,7 +10,6 @@ function MainRouter() {
       <Switch>
         <Route exact path={process.env.PUBLIC_URL + '/'}>
           <Collections />
-          <Users />
         </Route>
         <Route path='*'>
           <Notfound />
